+++ src/ic-cdk/src/api/management_canister/main/types.rs
@@ -4,6 +4,23 @@ use serde::{Deserialize, Serialize};
 /// Canister ID is Principal.
 pub type CanisterId = Principal;
 
+/// todo
+#[derive(
+    CandidType, Serialize, Deserialize, Debug, PartialEq, Eq, PartialOrd, Ord, Hash, Clone, Default,
+)]
+pub enum LogVisibility {
+    #[default]
+    #[serde(rename = "controllers")]
+    /// Only controllers of the canister can access the logs.
+    Controllers,
+    #[serde(rename = "public")]
+    /// Everyone is allowed to access the canister's logs.
+    Public,
+    #[serde(rename = "allowed_viewers")]
+    /// Canister logs are visible to a set of principals.
+    AllowedViewers(Vec<Principal>),
+}
+
 /// Canister settings.
 ///
 /// The settings are optional. If they are not explicitly set, the default values will be applied automatically.
@@ -52,6 +69,10 @@ pub struct CanisterSettings {
     ///
     /// Default value: 5_000_000_000_000 (5 trillion cycles).
     pub reserved_cycles_limit: Option<Nat>,
+    /// Defines who is allowed to read the canister's logs.
+    ///
+    /// Default value: Controllers
+    pub log_visibility: Option<LogVisibility>,
     /// Must be a number between 0 and 2<sup>48</sup>-1 (i.e 256TB), inclusively.
     ///
     /// It indicates the upper limit on the WASM heap memory consumption of the canister.
@@ -317,6 +338,8 @@ pub struct DefiniteCanisterSettings {
     pub freezing_threshold: Nat,
     /// Reserved cycles limit.
     pub reserved_cycles_limit: Nat,
+    /// Visibility of canister logs.
+    pub log_visibility: LogVisibility,
     /// The Wasm memory limit.
     pub wasm_memory_limit: Nat,
 }
@@ -434,6 +457,19 @@ pub struct CodeDeploymentRecord {
     pub module_hash: Vec<u8>,
 }
 
+/// Details about loading canister snapshot.
+#[derive(
+    CandidType, Serialize, Deserialize, Debug, PartialEq, Eq, PartialOrd, Ord, Hash, Clone,
+)]
+pub struct LoadSnapshotRecord {
+    /// The version of the canister at the time that the snapshot was taken
+    pub canister_version: u64,
+    /// The ID of the snapshot that was loaded.
+    pub snapshot_id: SnapshotId,
+    /// The timestamp at which the snapshot was taken.
+    pub taken_at_timestamp: u64,
+}
+
 /// Details about updating canister controllers.
 #[derive(
     CandidType, Serialize, Deserialize, Debug, PartialEq, Eq, PartialOrd, Ord, Hash, Clone,
@@ -457,6 +493,9 @@ pub enum CanisterChangeDetails {
     /// See [CodeDeploymentRecord].
     #[serde(rename = "code_deployment")]
     CodeDeployment(CodeDeploymentRecord),
+    /// See [LoadSnapshotRecord].
+    #[serde(rename = "load_snapshot")]
+    LoadSnapshot(LoadSnapshotRecord),
     /// See [ControllersChangeRecord].
     #[serde(rename = "controllers_change")]
     ControllersChange(ControllersChangeRecord),
@@ -505,4 +544,57 @@ pub struct CanisterInfoResponse {
     pub module_hash: Option<Vec<u8>>,
     /// Controllers of the canister.
     pub controllers: Vec<Principal>,
+}
+
+/// ID of a canister snapshot.
+pub type SnapshotId = Vec<u8>;
+
+/// A snapshot of the state of the canister at a given point in time.
+#[derive(
+    CandidType, Serialize, Deserialize, Debug, PartialEq, Eq, PartialOrd, Ord, Hash, Clone, Default,
+)]
+pub struct Snapshot {
+    /// ID of the snapshot.
+    pub id: SnapshotId,
+    /// The timestamp at which the snapshot was taken.
+    pub taken_at_timestamp: u64,
+    /// The size of the snapshot in bytes.
+    pub total_size: u64,
+}
+
+/// Argument type of [take_canister_snapshot](super::take_canister_snapshot).
+#[derive(
+    CandidType, Serialize, Deserialize, Debug, PartialEq, Eq, PartialOrd, Ord, Hash, Clone,
+)]
+pub struct TakeCanisterSnapshotArgs {
+    /// Principal of the canister.
+    pub canister_id: CanisterId,
+    /// An optional snapshot ID to be replaced by the new snapshot.
+    ///
+    /// The snapshot identified by the specified ID will be deleted once a new snapshot has been successfully created.
+    pub replace_snapshot: Option<SnapshotId>,
+}
+
+/// Argument type of [load_canister_snapshot](super::load_canister_snapshot).
+#[derive(
+    CandidType, Serialize, Deserialize, Debug, PartialEq, Eq, PartialOrd, Ord, Hash, Clone,
+)]
+pub struct LoadCanisterSnapshotArgs {
+    /// Principal of the canister.
+    pub canister_id: CanisterId,
+    /// ID of the snapshot to be loaded.
+    pub snapshot_id: SnapshotId,
+    /// sender_canister_version must be set to ic_cdk::api::canister_version().
+    pub sender_canister_version: Option<u64>,
+}
+
+/// Argument type of [delete_canister_snapshot](super::delete_canister_snapshot).
+#[derive(
+    CandidType, Serialize, Deserialize, Debug, PartialEq, Eq, PartialOrd, Ord, Hash, Clone,
+)]
+pub struct DeleteCanisterSnapshotArgs {
+    /// Principal of the canister.
+    pub canister_id: CanisterId,
+    /// ID of the snapshot to be deleted.
+    pub snapshot_id: SnapshotId,
 }