+++ src/ic-cdk/CHANGELOG.md
@@ -6,19 +6,48 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Changed
+
+- Add `AllowedViewers` variant to `LogVisibility` enum. (#512)
+
+### Added
+
+- Support Threshold Schnorr signing management canister API. (#518)
+
+## [0.16.0] - 2024-08-27
+
+### Changed
+
+- BREAKING: Add the `LoadSnapshot` variant to `CanisterChangeDetails`. (#504)
+
+### Added
+
+- Support Canister State Snapshots. (#504)
+  - Add methods: `take_canister_snapshot`, `load_canister_snapshot`, `list_canister_snapshots`, `delete_canister_snapshot`
+  - Add types: `LoadSnapshotRecord`, `SnapshotId`, `Snapshot`, `TakeCanisterSnapshotArgs`, `LoadCanisterSnapshotArgs`, `DeleteCanisterSnapshotArgs`
+
+## [0.15.0] - 2024-07-01
+
+### Changed
+
+- BREAKING: Stable Memory always use 64-bit addresses and `stable64_*` system API. (#498)
+- BREAKING: Add `log_visibility` to the management canister API types: (#497)
+   - `CanisterSettings`
+   - `DefiniteCanisterSettings`.
+
 ## [0.14.0] - 2024-05-17
 ## [0.13.3] - 2024-05-10 (yanked)
 
 ### Added
 
-- Add `wasm_memory_limit` to the management canister API types: (#483)
-   * `CanisterSettings`
-   * `DefiniteCanisterSettings`.
 - Provide safe wrapper of `in_replicated_execution` in ic-cdk. (#489)
 
 ### Changed
 
 - Upgrade `ic0` to v0.23.0. (#489)
+- BREAKING: Add `wasm_memory_limit` to the management canister API types: (#483)
+   - `CanisterSettings`
+   - `DefiniteCanisterSettings`.
 
 ## [0.13.2] - 2024-04-08